MINOR: Time and log producer state recovery phases (#10241)
During a slow log recovery it's easy to think that loading .snapshot files is a multi-second process. Often it isn't the snapshot loading that takes most of the time, rather it's the time taken to further rebuild the producer state from segment files. This patch times both snapshot load and segment recovery phases to better indicate what is taking time.

Reviewers: David Jacot <djacot@confluent.io>