KAFKA-9189: Use MetadataCache instead of ZK during controlled shutdown to avoid hang (#10361)
This avoids hanging during shutdown if ZK is unavailable. We could change ZK
calls to get the controller id and the broker information to have a timeout, but I
think this approach is better.

The downside is that the metadata cache may be slightly out of date, but we will
retry as per the controlled shutdown configuration. If this broker is partitioned
away from the Controller and is not receiving metadata updates, then we want
to shutdown asap anyway.

I added a test that timed out without this change and included a couple of clean-ups
in `ServerShutdownTest`:
* Removed `testCleanShutdownWithDeleteTopicEnabled`, which is redundant
since delete topics is enabled by default.
* Removed redundant method arguments

Reviewers: David Jacot <djacot@confluent.io>, Jun Rao <junrao@gmail.com>